+++ vllm/compilation/passes/pass_manager.py
@@ -17,7 +17,7 @@ from vllm.utils.system_utils import set_env_var
 from .vllm_inductor_pass import VllmInductorPass
 
 if rocm_aiter_ops.is_enabled():
-    from .fusion.rocm_aiter_allreduce_fusion import RocmAiterAllReduceFusionPass
+    from .fusion.allreduce_rms_fusion import RocmAiterAllReduceFusionPass
     from .fusion.rocm_aiter_fusion import (
         RocmAiterRMSNormQuantFusionPass,
         RocmAiterSiluMulFp8GroupQuantFusionPass,
